requirements.txt
# Core dependencies for LLM training on M3 MacBook Pro
# PyTorch with MPS support (Metal Performance Shaders for M3)
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Hugging Face ecosystem
transformers>=4.36.0
datasets>=2.16.0
tokenizers>=0.15.0
accelerate>=0.25.0
peft>=0.7.0  # Parameter-Efficient Fine-Tuning (LoRA, QLoRA)

# Data processing
pandas>=2.1.0
numpy>=1.24.0
pyarrow>=14.0.0
markdown>=3.5.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Monitoring and visualization
tensorboard>=2.15.0
tqdm>=4.66.0
wandb>=0.16.0  # Optional, can be disabled

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=1.0.0
psutil>=5.9.0  # For memory monitoring

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.12.0
flake8>=6.1.0
mypy>=1.7.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.5.0
